Internal: Move promise.py into modules subdirectory.
The promise is an internal module without any interfaces required by Sublime Text.
Therefore it is moved to a sub directory to keep it private.
